Switch cache functional tests to the new container support code in salt-factories
Additionally, split the tests into their own modules and clean up unused, or
no longer necessary code.
