+++ tests/pytests/functional/conftest.py
@@ -2,16 +2,7 @@ import logging
 import shutil
 
 import pytest
-from saltfactories.daemons.container import Container
 from saltfactories.utils.functional import Loaders
-
-try:
-    import docker
-except ImportError:
-    # Test suites depending on docker should be using
-    #     docker = pytest.importorskip("docker")
-    # so any fixtures using docker shouldn't ever be called or used.
-    pass
 
 log = logging.getLogger(__name__)
 
@@ -93,104 +84,3 @@ def reset_loaders_state(loaders):
     finally:
         # Reset the loaders state
         loaders.reset_state()
-
-
-@pytest.fixture(scope="module")
-def docker_client():
-    try:
-        client = docker.from_env()
-    except docker.errors.DockerException:
-        pytest.skip("Failed to get a connection to docker running on the system")
-    connectable = Container.client_connectable(client)
-    if connectable is not True:  # pragma: nocover
-        pytest.skip(connectable)
-    return client
-
-
-def pull_or_skip(image_name, docker_client):
-    try:
-        docker_client.images.pull(image_name)
-    except docker.errors.APIError as exc:
-        pytest.skip("Failed to pull docker image {!r}: {}".format(image_name, exc))
-    except ImportError:
-        pytest.skip("docker module was not installed")
-    return image_name
-
-
-@pytest.fixture(scope="module")
-def docker_redis_image(docker_client):
-    return pull_or_skip("redis:alpine", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_consul_image(docker_client):
-    return pull_or_skip("consul:latest", docker_client)
-
-
-# Pytest does not have the ability to parametrize fixtures with parametriezed
-# fixtures, which is super annoying. In other words, in order to have a `cache`
-# test fixture that takes different versions of the cache that depend on
-# different docker images, I've gotta make up fixtures for each
-# image+container. When https://github.com/pytest-dev/pytest/issues/349 is
-# actually fixed then we can go ahead and refactor all of these mysql images
-# (and container fixtures depending on it) into a single fixture.
-
-
-@pytest.fixture(scope="module")
-def docker_mysql_5_6_image(docker_client):
-    return pull_or_skip("mysql/mysql-server:5.6", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mysql_5_7_image(docker_client):
-    return pull_or_skip("mysql/mysql-server:5.7", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mysql_8_0_image(docker_client):
-    return pull_or_skip("mysql/mysql-server:8.0", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mariadb_10_1_image(docker_client):
-    return pull_or_skip("mariadb:10.1", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mariadb_10_2_image(docker_client):
-    return pull_or_skip("mariadb:10.2", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mariadb_10_3_image(docker_client):
-    return pull_or_skip("mariadb:10.3", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mariadb_10_4_image(docker_client):
-    return pull_or_skip("mariadb:10.4", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_mariadb_10_5_image(docker_client):
-    return pull_or_skip("mariadb:10.5", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_percona_5_5_image(docker_client):
-    return pull_or_skip("percona:5.5", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_percona_5_6_image(docker_client):
-    return pull_or_skip("percona:5.6", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_percona_5_7_image(docker_client):
-    return pull_or_skip("percona:5.7", docker_client)
-
-
-@pytest.fixture(scope="module")
-def docker_percona_8_0_image(docker_client):
-    return pull_or_skip("percona:8.0", docker_client)
