+++ tests/unit/test_module_names.py
@@ -47,6 +47,7 @@ EXCLUDED_FILES = [
     os.path.join("tests", "virtualname.py"),
     os.path.join("tests", "wheeltest.py"),
     os.path.join("tests", "zypp_plugin.py"),
+    os.path.join("tests", "pytests", "functional", "cache", "helpers.py"),
     os.path.join("tests", "pytests", "unit", "states", "virt", "helpers.py"),
 ]
